requirements.txt
flake8>=3.8.3
flake8-blind-except
flake8-builtins
flake8-eradicate
flake8-import-order
jsonschema>=3.2.0
jwcrypto
kafka-python==2.0.2
pep8-naming
py
pytest==7.4.4
pytest-html
pytest-instrument==0.3.1
pytest-lazy-fixture
pytest-repeat
pytest-split
pytest-timeout
pytest-rerunfailures
pytest-split
pyyaml>=5.3.1
requests
requests[socks]
pysocks
pytest-random-order
pytest-xdist
six
schema
selenium>=4.0.0
vulture
xmlschema==1.10.0
xmltodict
xlrd
pandas